requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
httpx==0.24.0
alembic~=1.15.2
python-dotenv==1.0.0
pydantic~=1.10.21
PyJWT==2.8.0